backend/requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Bcrypt==1.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==0.21.1
requests==2.28.2
tgtg==0.14.0
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.13.0
flask-bcrypt==1.0.1
python-dotenv==0.21.1
flask_bcrypt==1.0.1